kernel/process: switch ProcessId to use u64 unique identifier
This switches the unique numeric identifier of a process instance within the
ProcessId type (colloquially also called process ID) to be an 8-byte unsigned
integer, instead of usize. Additionally, it changes the kernel to panic on the
offchance that the counter used for assigning new process IDs rolls over.

This change is motivated by recent discussions around new IPC mechanisms. These
mechanisms consist of a "discovery" phase, where applications search for, and
potentially authenticate a given process, and a "communication" phase, where
they actually exchange data between processes. The discovery phase emits a
unique process handle that a subsequent communication phase will use to uniquely
identify a previously discovered and authenticated process. Virtually the only
good candidate for such a handle on a Tock system is the numeric process ID, as
it refers to a particular instance of an application, and does not require us to
keep track of extra state or mappings.

However, depending on which architecture we are running on, this numeric process
ID will be generated based on a counter that is either 4 or 8 byte
wide (`usize`). While Tock systems generally are limited in the number of
applications they can concurrently run, this identifier instead refers to an
_instance_ of a process. As such, this counter can be controlled and increased
by processes themselves. If this counter is based on a 4 byte value it is
unlikely, but not impossible, that malicious or colluding processes could force
it to overflow to a known, previous assigned value within a somewhat reasonable
timeframe (multiple days).

We should not prevent such overflows by panicing the kernel, or preventing
process restarts, as that endangers availability of the system, esp. when
deployed in critical and long running scenarios.

At the same time, while unlikely, explicitly permitting counter overflows forces
us to think about, model, and either handle or explicitly ignore many types of
attacks (e.g., confused deputy) that could arise from such a forced counter
overflow. This is not just affecting IPC, but any kernel or userspace subsystem
that explicitly or implicitly relies on process ID uniqueness.

This instead proposes to make this numeric counter consistent across platforms
and debug/release builds: it is always an 8 byte unsigned integer, and in
constrast to the previous implementation (which only panics on debug builds) an
overflow of this counter always panics the kernel. However, given the
astoronomical size of this counter, it is virtually impossible that a correct
kernel implementation will ever reach this case.

This does come with some drawbacks: I don't believe permformance will be
impacted significantly (comparision and increments of 64 bit numbers is decently
efficient even on 32 bit systems). However, it does add a single word of memory
everywhere that a Process ID is used. It also forces backwards-incompatible
changes to a few userspace APIs (legacy IPC being a notable example, although
we're in the process of removing that).

+++ kernel/src/process.rs
@@ -91,7 +91,7 @@ pub struct ProcessId {
     /// process at the given index does not match the value saved here, then the
     /// process moved or otherwise ended, and this `ProcessId` is no longer
     /// valid.
-    identifier: usize,
+    identifier: u64,
 }
 
 impl PartialEq for ProcessId {
@@ -128,7 +128,7 @@ impl fmt::Debug for ProcessId {
 impl ProcessId {
     /// Create a new `ProcessId` object based on the app identifier and its
     /// index in the processes array.
-    pub(crate) fn new(kernel: &'static Kernel, identifier: usize, index: usize) -> ProcessId {
+    pub(crate) fn new(kernel: &'static Kernel, identifier: u64, index: usize) -> ProcessId {
         ProcessId {
             kernel,
             index,
@@ -143,7 +143,7 @@ impl ProcessId {
     /// external implementations of `Process` can use it.
     pub fn new_external(
         kernel: &'static Kernel,
-        identifier: usize,
+        identifier: u64,
         index: usize,
         _capability: &dyn capabilities::ExternalProcessCapability,
     ) -> ProcessId {
@@ -168,7 +168,7 @@ impl ProcessId {
         }
     }
 
-    /// Get a `usize` unique identifier for the app this `ProcessId` refers to.
+    /// Get a `u64` unique identifier for the app this `ProcessId` refers to.
     ///
     /// This function should not generally be used, instead code should just use
     /// the `ProcessId` object itself to refer to various apps on the system.
@@ -182,7 +182,7 @@ impl ProcessId {
     /// the app may have restarted, or may have been ended or removed by the
     /// kernel. Therefore, calling `id()` is _not_ a valid way to check that an
     /// application still exists.
-    pub fn id(&self) -> usize {
+    pub fn id(&self) -> u64 {
         self.identifier
     }
